refactor: split G-code import page into widgets
