+++ lib/gcode_import/widgets/gcode_import_header.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class GCodeImportHeader extends StatelessWidget {
+  const GCodeImportHeader({super.key, required this.text});
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(text, style: Theme.of(context).textTheme.bodyMedium);
+  }
+}
